requirements.txt
# GPT MCP Server Dependencies
# ============================
# 설계: Kevin Zhang (DevOps Lead)

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Async I/O
aiofiles>=23.2.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.5.0

# Utilities
httpx>=0.25.0  # For testing

# Development (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
